requirements.txt
numpy==1.19.2
streamlit==0.67.1
pandas==1.0.5
matplotlib==3.3.0
seaborn==0.10.1
shap==0.37.0
plotly==4.14.3
scikit-learn==0.23.1
scipy==1.5.2
xgboost==1.3.3
bokeh==2.2.3
Pillow